+++ test/execution_test.ts
@@ -37,6 +37,7 @@ import {newJsonFetchResponse} from '../testing/mocks';
 import {newMockExecutionContext} from '../testing/mocks';
 import {newMockSyncExecutionContext} from '../testing/mocks';
 import {newPack} from '../builder';
+import path from 'path';
 import sinon from 'sinon';
 
 describe('Execution', () => {
@@ -444,10 +445,14 @@ describe('Execution', () => {
   });
 
   describe('CLI execution', () => {
+    let mockPrint: sinon.SinonStub;
     let mockPrintFull: sinon.SinonStub;
+    let mockExit: sinon.SinonStub;
 
     beforeEach(() => {
+      mockPrint = sinon.stub(helpers, 'print');
       mockPrintFull = sinon.stub(helpers, 'printFull');
+      mockExit = sinon.stub(process, 'exit');
     });
 
     afterEach(() => {
@@ -565,6 +570,42 @@ describe('Execution', () => {
           });
           const result = mockPrintFull.args[0][0];
           assert.deepEqual(result, [{value: 'foo', display: 'foo'}]);
+        });
+
+        it('pack exceptions print stacktrace', async () => {
+          await executeFormulaOrSyncFromCLI({
+            vm,
+            formulaName: 'Throw',
+            params: [],
+            manifest: fakePack,
+            manifestPath: '',
+            bundleSourceMapPath,
+            bundlePath,
+            contextOptions: {useRealFetcher: false},
+          });
+          const error = mockPrint.args[0][0];
+          const exitValue = mockExit.args[0][0];
+          assert.instanceOf(error, Error);
+          assert.include(error.stack, path.join(__dirname, 'packs/fake.ts'));
+          assert.equal(exitValue, 1);
+        });
+
+        it('CLI errors don\'t print stacktrace', async () => {
+          await executeFormulaOrSyncFromCLI({
+            vm,
+            formulaName: 'NotRealFormula',
+            params: [],
+            manifest: fakePack,
+            manifestPath: '',
+            bundleSourceMapPath,
+            bundlePath,
+            contextOptions: {useRealFetcher: false},
+          });
+          const message = mockPrint.args[0][0];
+          const exitValue = mockExit.args[0][0];
+          assert.typeOf(message, 'string');
+          assert.notInclude(message, path.join(__dirname, '/packs-sdk/dist/'));
+          assert.equal(exitValue, 1);
         });
       });
     }